Move ICU4C utils to devtools module
